week_2_workflow_orchestration/assignment/dockerfile-requirement.txt
pandas==1.5.2
prefect-gcp[cloud_storage]==0.2.5
protobuf==4.21.11
pyarrow==10.0.1
psycopg2-binary==2.9.5
sqlalchemy==1.4.46